[Fix] MIDI mapping: Pattern record checkbox wasn't updated properly. [Fix] Sample map: was broken when dealing with sample indexes greater than 255. [Mod] Version: Changed version number. [Ref] Minor code cleanup.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@380 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/dlg_misc.h
@@ -535,7 +535,7 @@ protected:
 	CSliderCtrl m_SbOctave;
 	CSoundFile *m_pSndFile;
 	UINT m_nInstrument;
-	BYTE KeyboardMap[NOTE_MAX];
+	WORD KeyboardMap[NOTE_MAX];
 
 public:
 	CSampleMapDlg(CSoundFile *pSndFile, UINT nInstr, CWnd *parent=NULL):CDialog(IDD_EDITSAMPLEMAP, parent)
